int of abs

+++ PCA9685_ArmBot/servoTest2.py
@@ -155,7 +155,7 @@ def move_to_angle(flexing_servo, starting_rotation, ending_rotation):
     flexing_rotation = starting_rotation
     print(abs((starting_rotation-ending_rotation)/2))
     
-    for each in range(abs((starting_rotation-ending_rotation)/2)):
+    for each in range(int(abs((starting_rotation-ending_rotation)/2))):
         if flexing_rotation > 177:
             flexing_rotation = 177
         # if flexing_rotation > 267:
